refactor: create componentes and change strategy of page
