+++ frontend/src/pages/UpdateProfile/index.js
@@ -1,0 +1,3 @@
+import UpdateProfile from './UpdateProfile'
+
+export { UpdateProfile }
