+++ frontend/src/pages/UpdateProfile/style.scss
@@ -1,4 +1,4 @@
-.containerUpdateEmail {
+.containerUpdate {
   padding: 1rem;
   &__group {
     margin-top: 1rem;
